Module miniz_oxide_c_api::tinfl

Modules

inflate_flags

Structs

tinfl_decompressor
Main decompression struct.

Constants

TINFL_DECOMPRESS_MEM_TO_MEM_FAILED

Functions

decompress
Main decompression function. Keeps decompressing data from in_buf until the in_buf is emtpy, out_cur is full, the end of the deflate stream is hit, or there is an error in the deflate stream.

tinfl_decompress⚠

tinfl_decompress_mem_to_heap⚠
Decompress data from p_src_buf to a continuously growing heap-allocated buffer.

tinfl_decompress_mem_to_mem⚠